Refactor and simplify node red nodes
Refactors the api node to using our API client instead of running the
request ourself. Cleans up the node, to only accept client-credentials
files as we don't expect anyone to need to overwrite any options, the
node is environment agnostic.

The insert node is cleaned up and simplified. Buffering is now handled
by a dedicated class, which makes the logic easier to reason about.
Reporting is also refactored to its own class which makes it easier to
follow.

We’re now relying on Joi for message validation.