Simplify Clarify Database class
The old database class has more complicated than needed, as it supported
multiple database connections per instance. The way node-red is designed
it would never happen—as any changes it settings would trigger a deploy,
which would re-create the node. Which meant it only cached one single
instance.

If the user were to create multiple `clarify_api` nodes for the same
integration, the database would be overwritten and any changes wouldn't
be synced. To mittigate part of the problem, the `node-id` is included
in the path, which means all metadata is contained to one `clarify_api`
node. Node-red reuses the same `clarify_api` node when creating multiple
`clarify_insert` nodes pointing to the api node.

Last I chagned to format of the file, from being an array of entries to
simply a map between the `input-id` and the hash, should speed up the
lookup.