Improve supported payloads
Supports sending in multiple `payload` formats. You're now allowed to
send in just a number as the payload or an object containing a time and
a value.

Support `valueType` which is the new name of the `type` property. We
still support the old one, both in the node-red plugin and on the API.

Reintroduce support for sending enumValues.

Moved the validation into its own file, which makes it easier to write
tests for.