Remove output from `clarify_insert` node
The output has never been defined and hasn't been very useful.
Previously all errors was moved to being thrown instead of returned as
an output. It made the output even less useful as inputs are buffered
and might yield two requests and two output calls. This seems to be the
most common pattern for other nodes as well (E.g. mqtt out).