Handle close event better
The Clarify API node now waits until all of it requests has finished
before closing itself. The Insert node sends all buffered data to the
insert node. As the API node lives in the global space, it will be
closed after the insert node, which means it will live for the lifespan
of the request.